requirements.txt
transformers
requests
gitpython
python-dotenv